Added form functionality
There is a rudimentary form functionality included in the project. You can make your own form and use it in a template or use a default one provided and style it using css.

To use the default, use @include('forms._show', App\Form::find($id)) in your blade file. It will load the given forms ID. You can find the ID in the admin panel, it's the last digit of the Action URL.

To build up your own form using HTML, open the form using the POST method and use the action in the admin panel. Don't forget to use the CSRF field in your form ( @csrf )