Bugfix for update meta-code
When saving a new meta, the code is automatically set to a sluggified label in the Meta's boot method.

This works fine, but conflicts with the MetaSeeder as that wasn't updated to the new boot yet.

This commit fixes that. It means that all references to 'level' and 'last_login' are updated to 'access-level' and 'last-login'.