Add the ability to submit spam
This commit adds the ability to submit spam from the admin panel.

If an Akismet key is available, it shows a flag on admin.comment.index.

Clicking the flag will submit the comment to Akismet and set the approved status to 0.

Also fixed a bug where an update on a comment will trigger a saved event, including the "new comment" flow.

Fixed a minor issue where the Custom class routes setting gives an error in the routes/web file when input is not as expected.